Reorganize Signing tests and CommandLine options tests
