+++ src/ImageBuilder.Tests/CommandLine/BuildOptionsBindingTests.cs
@@ -5,9 +5,9 @@
 using Microsoft.DotNet.ImageBuilder.Commands;
 using Shouldly;
 using Xunit;
-using static Microsoft.DotNet.ImageBuilder.Tests.Commands.OptionsBindingTestHelper;
+using static Microsoft.DotNet.ImageBuilder.Tests.CommandLine.OptionsBindingTestHelper;
 
-namespace Microsoft.DotNet.ImageBuilder.Tests.Commands;
+namespace Microsoft.DotNet.ImageBuilder.Tests.CommandLine;
 
 public class BuildOptionsBindingTests
 {
